Move ExtractEntityFromMessage to MessageHelpers & small tweaks
